rename: 기존의 myPage 를 myInfoPage로 변경
