+++ src/router/index.js
@@ -1,12 +1,14 @@
 // router/index.js 또는 router.js
 import { createRouter, createWebHistory } from "vue-router";
+import KakaoCallbackPage from "@/pages/auth/KakaoCallbackPage.vue";
 
 /* ── 페이지 컴포넌트 ─────────────────────── */
 import HomePage from "@/pages/HomePage.vue";
-import MyPage from "@/pages/user/MyPage.vue";
+import MyInfoPage from "@/pages/user/MyInfoPage.vue";
 import RecommendDeposit from "@/pages/deposit/RecommendDeposit.vue";
 import SignupPage from "@/pages/auth/SignupPage.vue";
 import LoginPage from "@/pages/auth/LoginPage.vue";
+import ResetPasswordPage from "@/pages/auth/ResetPasswordPage.vue";
 import UpdateUserInfoPage from "@/pages/user/UpdateUserInfoPage.vue";
 import UpdatePasswordPage from "@/pages/user/UpdatePasswordPage.vue";
 
@@ -30,27 +32,32 @@ const routes = [
   {
     path: "/oauth/kakao/callback",
     name: "KakaoCallback",
-    component: () => import("@/pages/auth/KakaoCallbackPage.vue"),
+    component: KakaoCallbackPage,
   },
-  { path: "/mypage", name: "mypage", component: MyPage },
   {
-    path: "/mypage/update",
+    path: "/reset-password",
+    name: "resetPassword",
+    component: ResetPasswordPage,
+  },
+  { path: "/myInfoPage", name: "myInfoPage", component: MyInfoPage },
+  {
+    path: "/myInfoPage/update",
     name: "updateUserInfo",
     component: UpdateUserInfoPage,
   },
   {
-    path: "/mypage/update/password",
+    path: "/myInfoPage/update/password",
     name: "updatePassword",
     component: UpdatePasswordPage,
   },
 
   {
-    path: "/mypage/update",
+    path: "/myInfoPage/update",
     name: "updateUserInfo",
     component: UpdateUserInfoPage,
   },
   {
-    path: "/mypage/update/password",
+    path: "/myInfoPage/update/password",
     name: "updatePassword",
     component: UpdatePasswordPage,
   },
